opensource readiness: rename gem
renames the gem from `trmnl_framework` to `trmnl-framework`
