+++ lib/trmnl-framework.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# Bundler auto-requires a gem by its package name (`trmnl-framework`). Hyphenated
+# gem names map to this entry file; the real load path is nested under trmnl/.
+require "trmnl/framework"
