+++ lib/trmnl/framework/static.rb
