+++ package.json
@@ -2,7 +2,7 @@
   "name": "trmnl-framework",
   "version": "3.2.0",
   "private": true,
-  "description": "Build tooling for the TRMNL ePaper design framework: CSS/JS minifiers and the Playwright runtime and visual suites. The framework itself ships as the trmnl_framework Ruby gem.",
+  "description": "Build tooling for the TRMNL ePaper design framework: CSS/JS minifiers and the Playwright runtime and visual suites. The framework itself ships as the trmnl-framework Ruby gem.",
   "license": "MIT",
   "repository": {
     "type": "git",
